Decrease McCabe complexity of concordance methods Break up concordance and BTree from utils.py into separate files
